pyproject.toml
[project]
name = "suguru"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiosqlite>=0.21.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "greenlet>=3.2.2",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]

